csc330projectdemo-main/requirements.txt
flask
flask-sqlalchemy
flask-wtf
python-dotenv
email-validator
